Clean up docs, move discard tests to their own file
